+++ src/main/java/com/elmakers/mine/bukkit/entity/EntityData.java
@@ -122,6 +122,8 @@ public class EntityData implements com.elmakers.mine.bukkit.api.entity.EntityDat
     protected LinkedList<WeightedPair<String>> spells;
     protected boolean requiresTarget;
     
+    protected ConfigurationSection disguise;
+    
     public EntityData(Entity entity) {
         this(entity.getLocation(), entity);
     }
@@ -243,6 +245,9 @@ public class EntityData implements com.elmakers.mine.bukkit.api.entity.EntityDat
             }
         }
 
+        disguise = ConfigurationUtils.getConfigurationSection(parameters, "disguise");
+
+        isBaby = parameters.getBoolean("baby", false);
         tickInterval = parameters.getLong("cast_interval", 0);
         if (parameters.contains("cast")) {
             spells = new LinkedList<WeightedPair<String>>();
@@ -625,6 +630,12 @@ public class EntityData implements com.elmakers.mine.bukkit.api.entity.EntityDat
             Mage apiMage = controller.getMage(entity);
             if (apiMage instanceof com.elmakers.mine.bukkit.magic.Mage) {
                 ((com.elmakers.mine.bukkit.magic.Mage)apiMage).setEntityData(this);
+            }
+        }
+        
+        if (controller != null && disguise != null) {
+            if (!controller.disguise(entity, disguise)) {
+                controller.getLogger().warning("Invalid disguise type: " + disguise.getString("type"));
             }
         }