+++ src/main/java/com/elmakers/mine/bukkit/magic/MagicController.java
@@ -787,7 +787,7 @@ public class MagicController implements MageController {
             libsDisguiseManager = new LibsDisguiseManager(plugin, libsDisguisePlugin);
             if (libsDisguiseManager.initialize()) {
                 if (libsDisguiseEnabled) {
-                    getLogger().info("Integrated with LibsDisguises, most spells can't be cast while disguised");
+                    getLogger().info("Integrated with LibsDisguises, mob disguises and disguise_restricted features enabled");
                 } else {
                     getLogger().info("LibsDisguises integration disabled");
                 }
@@ -4312,6 +4312,14 @@ public class MagicController implements MageController {
     @Override
     public boolean isDisguised(Entity entity) {
         return !libsDisguiseEnabled || libsDisguiseManager == null || entity == null ? false : libsDisguiseManager.isDisguised(entity);
+    }
+    
+    @Override
+    public boolean disguise(Entity entity, ConfigurationSection configuration) {
+        if (!libsDisguiseEnabled || libsDisguiseManager == null || entity == null) {
+            return false;
+        }
+        return libsDisguiseManager.disguise(entity, configuration);
     }
 
     @Override